modifying different AST into single files
