--- group2JavaCompiler/group2JavaCompiler/AST/AssignExpression.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace group2JavaCompiler
-{
-    class AssignExpression : Expression
-    {
-        private Expression lhs, rhs;
-
-        public AssignExpression(Expression lhs, Expression rhs)
-        {
-            this.lhs = lhs;
-            this.rhs = rhs;
-        }
-
-        public override void dump(int indent)
-        {
-            label(indent, "AssignmentExpression\n");
-            lhs.dump(indent + 1, "lhs");
-            rhs.dump(indent + 1, "rhs");
-        }
-    }
-}
