--- group2JavaCompiler/group2JavaCompiler/AST/IntegerLiteralExpression.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace group2JavaCompiler
-{
-    class IntegerLiteralExpression : Expression
-    {
-        private int value;
-
-        public IntegerLiteralExpression(int value)
-        {
-            this.value = value;
-        }
-
-        public override void dump(int indent)
-        {
-            label(indent, "IntegerLiteralExpression {0}\n", value);
-        }
-    }
-}
