+++ group2JavaCompiler/group2JavaCompiler/AST/Method.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace group2JavaCompiler
 {
-    class Method : Node
+    public class Method : Node
     {
         string modifier, name;
         Types type;
