+++ group2JavaCompiler/group2JavaCompiler/Program.cs
@@ -1,15 +1,21 @@
-﻿
+﻿using System.IO;
 namespace group2JavaCompiler
 {
     class Program
     {
         static void Main(string[] args)
         {
-          /*  var file = new System.IO.FileStream(args[0], System.IO.FileMode.Open);
-           javaCompilerScanner scanner = new Scanner(file);
-            //javaCompiler.Lexer.Scanner scanner = new javaCompiler.Lexer.Scanner(System.Console.OpenStandardInput());
+            var filename = args[0];
+            var file = new FileStream(args[0], FileMode.Open);
+            Scanner scanner = new Scanner(file);
             Parser parser = new Parser(scanner);
-            parser.Parse();*/
+            parser.Parse();
+            Parser.root.dump(0);
+            /*  var file = new System.IO.FileStream(args[0], System.IO.FileMode.Open);
+             javaCompilerScanner scanner = new Scanner(file);
+              //javaCompiler.Lexer.Scanner scanner = new javaCompiler.Lexer.Scanner(System.Console.OpenStandardInput());
+              Parser parser = new Parser(scanner);
+              parser.Parse();*/
 
             /*
             Statement root = new ExpressionStatement(new AssignExpression(new IdentifierExpression("x"), new IntegerLiteralExpression(42)));
@@ -24,15 +30,15 @@ namespace group2JavaCompiler
                       );
             */
 
-            Class root = new Class("public", "Test",
-                            new Method("public static",
-                            new NamedType("void"),
-                            "main",
-                            new Arguments(new ArrayType(new NamedType("String")), "args"),
-                            new VariableDeclarationStatement(new NamedType("int"), new IdentifierExpression("x"), new IntegerLiteralExpression(42))
-                            )
-                         );
-            root.dump(0);
+            /* Class root = new Class("public", "Test",
+                             new Method("public static",
+                             new NamedType("void"),
+                             "main",
+                             new Arguments(new ArrayType(new NamedType("String")), "args"),
+                             new VariableDeclarationStatement(new NamedType("double"), new IdentifierExpression("x"), new IntegerLiteralExpression(42))
+                             )
+                          );
+             root.dump(0);*/
         }
     }
 }
