+++ test/TestIfStatement.java
@@ -1,4 +1,4 @@
-public class test{
+public class test  final{
     public static void mains(String[] args)
     {
         if()
